Music systems implementation
Added three new scripts: MusicManager, MusicTrigger, and MusicSceneStart.
MusicManager is the master script for managing what music track is playing from the MainScene.
MusicTrigger can be attached to GameObjects to change what music track is playing when the player collides with their Collider2D set as trigger.
MusicSceneStart, if found in a scene, changes the track when a scene begins.